ctrl+alt+l

+++ src/main/java/com/github/ciomarabanu/leetcode/CombinationIterator.java
@@ -13,23 +13,23 @@ public class CombinationIterator {
         chars = characters;
         combLength = combinationLength;
         var asciiChars = new int[chars.length()];
-        for (int i = 0; i < chars.length(); i++){
-            asciiChars[i] = (int) chars.charAt(i);
+        for (int i = 0; i < chars.length(); i++) {
+            asciiChars[i] = chars.charAt(i);
         }
         sortedChars = Arrays.stream(asciiChars).sorted().toArray();
     }
 
     public String next() {
-        if (counter + combLength == sortedChars.length){
+        if (counter + combLength == sortedChars.length) {
             return "";
         }
         StringBuilder next = new StringBuilder(Character.toString(sortedChars[counter]));
-        if (nextJ + combLength == sortedChars.length){
+        if (nextJ + combLength == sortedChars.length) {
             counter++;
         }
-            for (int j = nextJ; j < combLength; j++){
-                next.append(Character.toString(sortedChars[j]));
-            }
+        for (int j = nextJ; j < combLength; j++) {
+            next.append(Character.toString(sortedChars[j]));
+        }
         nextJ++;
         return next.toString();
     }
